Revert pyside6 and shiboken6 pin version to 6.9
The snap uses the kde-neon-6 extension, which provides the Qt libraries. At some point yesterday, that extension's `stable` channel was inadvertently upgraded to Qt 6.10, which broke FreeCAD's snap build (which uses Qt 6.9). The FreeCAD snap's Qt version was pinned to 6.10 and rebuilt to 6.10 to fix this, which worked.

Yet this morning Ithe KDE folks have downgraded the kde-neon-6 extension dependencies back to 6.9, probably after realizing their mistake. This reverts the previous fix for the FreeCAD snap.